revisions API users: use release_revisions() in submodule.c edge case
Use release_revisions() on the the "struct rev_info" in
show_submodule_diff_summary() where we'd otherwise need to do the
equivalent of pre-initializing the "struct rev_info" with "{ 0 }" if
we were going to add it to the cleanup being performed in the "out"
part of the function, let's instead introduce an "out_no_rev" for the
reasons discussed in the preceding commit.

Doing so for the "goto" on "(!left || !right || !sub)" added in
8e6df65015f (submodule: refactor show_submodule_summary with helper
function, 2016-08-31) would have been straightforward, as in the
preceding commit.

But for the case of prepare_submodule_diff_summary() failing it's less
straightforward. Reading the pre-image we could simply retain the
"goto out" if it fails, because we can see that the first thing it
does is call repo_init_revisions().

But having a hard reliance on that would be a bit nasty, as we'd
potentially introduce a segfault if the function did some other
initialization first, and early aborted if that failed.

Let's just fold that helper function away into
show_submodule_diff_summary() itself, which was its only user. Now
following the flow of initialization is more obvious, and it's
immediately clear that the "goto out" if prepare_revision_walk()
returns non-zero is safe.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>